pyproject.toml
[project]
name = "vector-db-benchmark"
version = "0.1.0"
description = "OpenSearch vs Qdrant vector search benchmark using Upstage Embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opensearch-py>=2.4.0",
    "qdrant-client>=1.7.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "make>=0.1.6.post2",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "streamlit>=1.35.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
